# openapi/v1.yaml
...
components:
  schemas:
...
    OperationView:
      description: 'Wire shape for one operation.


        An explicit DTO rather than serialising the dataclass directly: the

        response schema is enforced field by field, so a future field added to

        :class:`Operation` for internal use cannot silently start being published

        (see the output-validation rule — an unvalidated response is as dangerous

        as an unvalidated request, just harder to notice).'
      properties:
        completed_at:
          anyOf:
            - type: string
            - type: 'null'
          title: Completed At
        created_at:
          anyOf:
            - type: string
            - type: 'null'
          title: Created At
        detail:
          anyOf:
            - type: string
            - type: 'null'
          title: Detail
        error:
          anyOf:
            - type: string
            - type: 'null'
          title: Error
        id:
          title: Id
          type: string
        is_terminal:
          title: Is Terminal
          type: boolean
        kind:
          title: Kind
          type: string
        progress:
          anyOf:
            - type: number
            - type: 'null'
          title: Progress
        resource_id:
          anyOf:
            - type: string
            - type: 'null'
          title: Resource Id
        resource_kind:
          anyOf:
            - type: string
            - type: 'null'
          title: Resource Kind
        result:
          anyOf:
            - additionalProperties: true
              type: object
            - type: 'null'
          title: Result
        state:
          title: State
          type: string
        status:
          title: Status
          type: string
        updated_at:
          anyOf:
            - type: string
            - type: 'null'
          title: Updated At
      required:
        - id
        - kind
        - state
        - status
        - is_terminal
        - resource_id
        - resource_kind
        - progress
        - detail
        - error
        - result
        - created_at
        - updated_at
        - completed_at
      title: OperationView
      type: object
...
  securitySchemes:
    bearerAuth:
      bearerFormat: JWT
      scheme: bearer
      type: http
info:
  title: PenguinCloud Portal API
  version: 1.0.0
openapi: 3.1.0
paths:
...
  /api/v1/products/{product_id}/operations:
    get:
      description: ''
      operationId: get_list_operations
      parameters:
        - in: path
          name: product_id
          required: true
          schema:
            type: number
      responses:
        '200':
          content:
            application/json:
              schema:
                description: Envelope for a page of operations.
                properties:
                  has_more:
                    title: Has More
                    type: boolean
                  operations:
                    items:
                      $ref: '#/components/schemas/OperationView'
                    maxItems: 10000
                    title: Operations
                    type: array
                  page:
                    title: Page
                    type: integer
                  per_page:
                    title: Per Page
                    type: integer
                  total:
                    anyOf:
                      - type: integer
                      - type: 'null'
                    title: Total
                required:
                  - operations
                  - page
                  - per_page
                  - total
                  - has_more
                title: OperationListResponse
                type: object
          description: Envelope for a page of operations.
      summary: List a product's long-running operations, newest first.
  /api/v1/products/{product_id}/operations/{kind}/{operation_id}:
    get:
      description: ''
      operationId: get_get_operation
      parameters:
        - in: path
          name: product_id
          required: true
          schema:
            type: number
        - in: path
          name: kind
          required: true
          schema:
            type: string
        - in: path
          name: operation_id
          required: true
          schema:
            type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                description: 'Wire shape for one operation.


                  An explicit DTO rather than serialising the dataclass directly: the

                  response schema is enforced field by field, so a future field added to

                  :class:`Operation` for internal use cannot silently start being published

                  (see the output-validation rule — an unvalidated response is as dangerous

                  as an unvalidated request, just harder to notice).'
                properties:
                  completed_at:
                    anyOf:
                      - type: string
                      - type: 'null'
                    title: Completed At
                  created_at:
                    anyOf:
                      - type: string
                      - type: 'null'
                    title: Created At
                  detail:
                    anyOf:
                      - type: string
                      - type: 'null'
                    title: Detail
                  error:
                    anyOf:
                      - type: string
                      - type: 'null'
                    title: Error
                  id:
                    title: Id
                    type: string
                  is_terminal:
                    title: Is Terminal
                    type: boolean
                  kind:
                    title: Kind
                    type: string
                  progress:
                    anyOf:
                      - type: number
                      - type: 'null'
                    title: Progress
                  resource_id:
                    anyOf:
                      - type: string
                      - type: 'null'
                    title: Resource Id
                  resource_kind:
                    anyOf:
                      - type: string
                      - type: 'null'
                    title: Resource Kind
                  result:
                    anyOf:
                      - additionalProperties: true
                        type: object
                      - type: 'null'
                    title: Result
                  state:
                    title: State
                    type: string
                  status:
                    title: Status
                    type: string
                  updated_at:
                    anyOf:
                      - type: string
                      - type: 'null'
                    title: Updated At
                required:
                  - id
                  - kind
                  - state
                  - status
                  - is_terminal
                  - resource_id
                  - resource_kind
                  - progress
                  - detail
                  - error
                  - result
                  - created_at
                  - updated_at
                  - completed_at
                title: OperationView
                type: object
          description: 'Wire shape for one operation.


            An explicit DTO rather than serialising the dataclass directly: the

            response schema is enforced field by field, so a future field added to

            :class:`Operation` for internal use cannot silently start being published

            (see the output-validation rule — an unvalidated response is as dangerous

            as an unvalidated request, just harder to notice).'
      summary: Poll one operation. This is the route the UI's refetch loop calls.
...
security:
  - bearerAuth: []